test: Fix up librados/neorados tests
There much mocking of exec going on in these tests. An extensive
search and replace was required to find them all.

Note that the failure looks like memory corruption when manipulating
the ops array - but the actually issue is the mocking not intercepting
the exec_impl call (and similar) correctly.

Signed-off-by: Alex Ainscow <aainscow@uk.ibm.com>