client: explicitly deny fscrypt fallocate
Do not allow fallocate if inode is encrypted. This operation if
implemented would require a first and/or last block RMW if offset
and/or offset+length is not aligned to fscrypt block size
(ie punch_hole). This sys call should not be modifying actual
data so we skip and do not support it.

This matches functionality from the kernel client:
https://github.com/torvalds/linux/blob/3f9f0252130e7dd60d41be0802bf58f6471c691d/fs/ceph/file.c#L2676-L2677

Fixes: https://tracker.ceph.com/issues/73497
Signed-off-by: Christopher Hoffman <choffman@redhat.com>