osd/scrub: support an operator-abort command
The new explicit command aborts any ongoing scrub of the target PG,
including operator-initiated scrubs. That additional capability is needed now that
operator-initiated scrubs are no longer blocked by 'no-scrub' settings.

The scenario we are trying to help the operator with is:
- an operator issues a set of operator-initiated scrubs (e.g., via a
  script), then realizes the mistake and wants to abort them all.
The abort command also downgrades the urgency level of the scrub target
(as otherwise the target would immediately restart, against the operator
wishes).

This commit implements the changes to the state machine and to the abort
logic, assuming the operator command was translated into an event.

Signed-off-by: Ronen Friedman <rfriedma@redhat.com>