src/cephadm: updated crash dir creation
This change is meant to address the issue of daemons like
grafana changing the uid and gid of the crash directory,
/var/lib/ceph<fsid>/crash, the crash service only works if
this directory has the standard ceph-user uid and gid. I've
added a condition which will only create a crash directory
if the crash daemon is being deployed. This ensures
uid and gid are always correct.

Signed-off-by: Timothy Q Nguyen <timqn22@gmail.com>