osd: write + truncate I/O sequence for ceph_test_rados_io_sequence
Test tool improvement.

Add a new I/O sequence 16 to ceph_test_rados_io_sequence which
tests permutations of writing and truncating an object. This gives more
coverage of the field found bug that was fixed by
https://tracker.ceph.com/issues/74128 but was not found in test. The
new test validates writing an object of size X and then truncating it
to size Y (Y may be smaller, the same or larger than X) and then
writing size X again for a variety of permutations of X and Y. In
particular this includes boundary conditions such as the end of a
4K page, end of a chunk and end of a stripe.

Signed-off-by: Bill Scales <bill_scales@uk.ibm.com>