client: mark write delegation as broken if it is recalled during a open() with read mode
nfs-ganesha sends a downgrade op to the client with the same fh after
performing a delegated write. This downgrade op calls ll_open() with read
mode, which calls Inode::get_open_ref -> Inode::break_deleg. Inside this
func ::break_deleg, a call to Inode::recall_deleg is made with skip_read=1,
i.e. to recall all the write delegations and skip recalling all the read
delegations. It leads to calling deleg.recall(1) where callback is invoked
recall_cb(fh, priv). This triggers ganesha to release the delegation and then
in the libcephfs, while (!delegations_broken(skip_read)) checks if the
delegations are broken, this code path is buggy. Inside if (skip_read), it
only checks for the read delegations, if its the case then the func returns
false meaning the delegations are not broken but it did break in the previous
call i.e. deleg.recall(). So, what happens next is client->wait_on_list
(waitfor_deleg); i.e. the client is waiting on the cond list forever.

Fixes: https://tracker.ceph.com/issues/74155
Signed-off-by: Dhairya Parmar <dparmar@redhat.com>