mgr/dashboard: add a command to refresh certificates
if you set a new certificate, you need to restart the dashboard module
for that certificate to get applied on the cherrypy. instead you just
need to reconfigure the ssl adapter for it to work.

reconfiguring adapter is the best way since it won't close the current
connection where the engine restart will close the connection which will
be an inconvenience.

apply the new certificates

```
ceph dashboard set-ssl-certificate -i <cert_file>
ceph dashboard set-ssl-certificate-key -i <key_file
```
and then run
```
ceph dashboard refresh-ssl-certificates
```

Signed-off-by: Nizamudeen A <nia@redhat.com>