crimson/os/seastore/lba_mapping: Simplify refresh interface
LBAMapping::refresh returned new LBAMapping instance
instead of reusing `this`.
The main reason was to avoid making sure that the callers keep
the LBAMapping instance alive while being refreshed.

The way the interface worked required creating a new mapping instance
each time a refresh() was needed - this largely complicated
the users logic.
This commit updates LBAMapping::refresh to reuse `this`.

* refresh is renamed to refresh_mapping for easier differentiation
  with LBACursor::refresh

* refresh would likely be deprecated soon, this PR would make this
  deprecation smoother once all users are one-liners instead of
  chained continuations.

Signed-off-by: Matan Breizman <mbreizma@redhat.com>