rbd-mirror: allow resync while a group snapshot is still syncing
currently we do not allow resync operation if the snapshot is still inprogress
to sync until its fully done. This means that if snapshot synchronization
becomes stuck for any reason, a resync cannot be triggered, resulting in an
undesirable operational limitation.

this change enables resync requests to be processed even when a group snapshot
is still syncing, allowing resync in the middle of syncing a group snapshot.

Note: this commit also brings a change to rbd_mirror_group_simple.sh
in the test_resync_marker() to fix an existing issue.

Signed-off-by: Prasanna Kumar Kalever <prasanna.kalever@redhat.com>