cleanup: simplify group image snapshot validation
validation of an image snapshot association with a group snapshot requires
checking either a valid group_spec or the presence of a group_snap_id
in cls::rbd::MirrorSnapshotNamespace.

this commit tries to remove redundant validation of checking for both and rely
on this minimal condition of checking for a valid group_spec.

Signed-off-by: Prasanna Kumar Kalever <prasanna.kalever@redhat.com>