rgw/multi: Reliably recover from future generation sync
Since we can't map shards between generations, if we get a sync entry
for a future shard, synthesize an entry on the current generation for
all shards.

To keep this from creating an unholy amount of work during reshard, we
add a cache dimensioned by bucket and generation to record the last
time we've done so, refraining if we have within the past hour.

Signed-off-by: Adam C. Emerson <aemerson@redhat.com>